refactor(frontend): extract community derived hook
